requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12

# ML and Embeddings
torch==2.6.0
torchvision==0.21.0
transformers==4.45.2
pillow==10.4.0
numpy==1.26.4
sentencepiece==0.2.0

# Vector Database
qdrant-client==1.12.0

# Cache and Queue
redis==5.1.1
celery==5.4.0

# Frontend MVP
streamlit==1.39.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.21.0
structlog==24.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
locust==2.32.0

# Utilities
python-dotenv==1.0.1
click==8.1.7
tqdm==4.66.5
pyyaml==6.0.2

# Development
black==24.10.0
ruff==0.6.9
mypy==1.11.2
ipykernel==6.29.5

# Notebooks and visualization (optional, for VS Code Interactive Window)
matplotlib==3.9.2
seaborn==0.13.2
umap-learn==0.5.6
scikit-learn==1.5.2
